+++ src/Options.js
@@ -1,0 +1,13 @@
+/**
+ * Class Options
+ */
+class Options {
+    constructor() {
+        /**
+         * @type {boolean}
+         */
+        this.experimental = false;
+    }
+}
+
+export {Options};